+++ src/scripts/wildBattle/RouteHelper.ts
@@ -1,3 +1,4 @@
+///<reference path="../../declarations/routes/Routes.d.ts"/>
 ///<reference path="../../declarations/routes/RoutePokemon.d.ts"/>
 
 /**
